feat(theme): update theme color
