+++ src/lib/styles/theme/components/index.ts
@@ -1,0 +1,7 @@
+import { Button } from "./button";
+import { Input } from "./input";
+
+export const components = {
+  Button,
+  Input,
+};
