+++ src/lib/styles/theme/components/input.ts
@@ -1,0 +1,11 @@
+import type { ComponentStyleConfig } from "@chakra-ui/react";
+
+export const Input: ComponentStyleConfig = {
+  sizes: {
+    lg: {
+      field: {
+        borderRadius: 24,
+      },
+    },
+  },
+};
