+++ src/lib/styles/theme/config.ts
@@ -1,0 +1,5 @@
+import type { ThemeConfig } from "@chakra-ui/react";
+
+export const config: ThemeConfig = {
+  disableTransitionOnChange: false,
+};
