+++ src/lib/styles/theme/index.ts
@@ -1,0 +1,15 @@
+import { extendTheme } from "@chakra-ui/react";
+
+import { colors } from "./colors";
+import { components } from "./components";
+import { config } from "./config";
+import { fonts } from "./fonts";
+
+const customTheme = extendTheme({
+  fonts,
+  colors,
+  components,
+  config,
+});
+
+export default customTheme;
