+++ src/pages/_app.tsx
@@ -1,12 +1,12 @@
 import { ChakraProvider } from "@chakra-ui/react";
-import { DefaultSeo } from "next-seo";
 import type { AppProps } from "next/app";
 import Head from "next/head";
+import { DefaultSeo } from "next-seo";
 import NextNProgress from "nextjs-progressbar";
 
 import defaultSEOConfig from "../../next-seo.config";
 import Layout from "lib/components/layout";
-import customTheme from "lib/styles/customTheme";
+import customTheme from "lib/styles/theme";
 import "lib/styles/globals.css";
 
 const MyApp = ({ Component, pageProps }: AppProps) => {
